Returns null when Content-Length > Integer.MAX_VALUE (#410)
OkHttp client should not bail out when response body is >2gb

Instead, we return null length and the response can be processed
similarly to a chunked body.

Also, shares common client tests in AbstractClientTest
