+++ httpclient/src/main/java/feign/httpclient/ApacheHttpClient.java
@@ -194,7 +194,8 @@ public final class ApacheHttpClient implements Client {
 
       @Override
       public Integer length() {
-        return entity.getContentLength() >= 0 ? (int) entity.getContentLength() : null;
+        return entity.getContentLength() >= 0 && entity.getContentLength() <= Integer.MAX_VALUE ?
+                (int) entity.getContentLength() : null;
       }
 
       @Override
